+++ examples/containers.rs
@@ -2,7 +2,7 @@ use gooey::value::Dynamic;
 use gooey::widget::{MakeWidget, WidgetInstance};
 use gooey::widgets::container::ContainerShadow;
 use gooey::window::ThemeMode;
-use gooey::{Gooey, Run};
+use gooey::Run;
 use kludgine::figures::units::Lp;
 use kludgine::figures::Point;
 
@@ -10,7 +10,7 @@ fn main() -> gooey::Result {
     let theme_mode = Dynamic::default();
     set_of_containers(3, theme_mode.clone())
         .centered()
-        .into_window(Gooey::default())
+        .into_window()
         .themed_mode(theme_mode)
         .run()
 }